requirements.txt
requests
beautifulsoup4
pandas
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0 